make token structs private
